reorganize login form
