+++ app/(auth)/login/donor/page.tsx
@@ -1,0 +1,11 @@
+import { LoginForm } from "@/src/components/page/login/login-form";
+
+export default function DonorLoginPage() {
+  return (
+    <div className="bg-muted flex min-h-svh flex-col items-center justify-center p-6 md:p-10">
+      <div className="w-full max-w-sm md:max-w-3xl">
+        <LoginForm />
+      </div>
+    </div>
+  );
+}
