+++ app/(auth)/login/organization/page.tsx
@@ -1,0 +1,11 @@
+import { OrganizationLoginForm } from "@/src/components/page/login/organization-login-form";
+
+export default function OrganizationLoginPage() {
+  return (
+    <div className="bg-muted flex min-h-svh flex-col items-center justify-center p-6 md:p-10">
+      <div className="w-full max-w-sm md:max-w-3xl">
+        <OrganizationLoginForm />
+      </div>
+    </div>
+  );
+}
